refactor(Validator): extract common function for lotto and bonus validator
