+++ src/main/kotlin/lotto/BonusValidator.kt
@@ -1,6 +1,12 @@
 package lotto
 
-interface BonusValidator {
+interface BonusValidator : Validator {
 
     fun validateNumber(number: Int)
+
+    fun validateDuplicatedNumber(number: Int, numbers: List<Int>) {
+        require(number !in numbers) {
+            "Bonus number must not duplicate with winning numbers."
+        }
+    }
 }
