+++ src/main/kotlin/lotto/LottoValidator.kt
@@ -1,10 +1,8 @@
 package lotto
 
-import lotto.Consts.END_NUMBER
 import lotto.Consts.LOTTO_NUMBER_SIZE
-import lotto.Consts.START_NUMBER
 
-interface LottoValidator {
+interface LottoValidator : Validator {
 
     fun validateNumbers(numbers: List<Int>) {
         numbers.forEach {
@@ -12,12 +10,6 @@ interface LottoValidator {
         }
         validateSize(numbers.size)
         validateDuplicatedNumber(numbers)
-    }
-
-    private fun validateInRange(number: Int) {
-        require(number in START_NUMBER..END_NUMBER) {
-            "Lotto number '$number' must be between $START_NUMBER and $END_NUMBER."
-        }
     }
 
     private fun validateSize(size: Int) {
